fix: apply orders-service code review fixes
Bugs: move setProcessID before persist(), null-guard promotion dates,
null-check product from API (new ProductNotFoundException), widen
count to long, Kafka bootstrap uses %dev/%prod profiles,
javax→jakarta.inject.Singleton.

Security: remove plaintext cardNumber/expirationDate from Order entity,
add @Authenticated to OrdersResource, add ownership checks on
deleteOrder/updateOrder/confirmOrder.

Architecture: replace banned quarkus-rest with quarkus-resteasy stack
(orders-service and both shared API poms), remove dual JSON provider
(quarkus-rest-jackson), drop dead @LoggingFilter.Logged on PricingApiClient,
move Kafka emit from OrderService to OrdersResource, switch
GlobalExceptionHandler to ExceptionMapper<Exception> with @Provider.

Style: remove System.out.println from LoggingFilter, unify
CartService.toResponse/toResponseWithPrices into one method,
rename ErrorResponse.errorCode to type.

Tests: add @BeforeEach Order.deleteAll() to OrdersResourceTest, annotate
each test with @TestSecurity matching the order owner, add getPrice
caching path tests to PriceCacheServiceTest, update CartResourceTest
assertions to new ErrorResponse shape.

+++ backend/orders-api/src/main/java/the/chak/ecommerce/orders/boundary/OrdersApi.java
@@ -20,7 +20,7 @@ public interface OrdersApi {
     @POST
     @Path("/search")
     @Produces(MediaType.APPLICATION_JSON)
-    Tuple<Integer, List<OrderDTO>> searchOrders(SearchOrdersCommand searchOrdersCommand);
+    Tuple<Long, List<OrderDTO>> searchOrders(SearchOrdersCommand searchOrdersCommand);
 
     @POST
     Response createOrder(OrderRequest orderRequest);
